% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{meta}
\alias{meta}
\title{Return meta Data
Get HCC risk-score related metaData.}
\usage{
meta()
}
\description{
Return meta Data
Get HCC risk-score related metaData.
}
\examples{
\dontrun{
meta()
metaDf = meta()
}
}
