% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{score
Accept a dataframe of HCC RNA data, and return 48 risk-scores in a new dataframe.}
\usage{
score(data)
}
\arguments{
\item{data}{RNA expression dataframe, samples as in column and gene Ensembl ID as row name.}
}
\description{
score
Accept a dataframe of HCC RNA data, and return 48 risk-scores in a new dataframe.
}
\section{Waining}{

The missing gene expression will be fill with 0.
}

\examples{
\dontrun{
df = read.delim("GEOdata.txt")
score_df = score(df)
}
}
